+++ CHANGELOG.md
@@ -18,6 +18,7 @@ nav_order: 1
 - Fix `aiven_kafka_quota` added retry logic to handle API eventual consistency
 - Improve `aiven_organization`: added retries to `AccountList` for eventual consistency
 - Improve `aiven_organization_project`: added retries to the read operation to ensure eventual consistency after creating or updating the resource
+- Fix `aiven_organization_application_user_token`: prevent `full_token` from being lost after running `terraform refresh`
 - Add `aiven_opensearch` field `opensearch_user_config.jwt`: OpenSearch JWT Configuration
 - Add `aiven_pg` field `pg_user_config.pg.io_combine_limit`: EXPERIMENTAL: Controls the largest I/O size in operations
   that combine I/O in 8kB units